+++ src/es2015/PlotBoilerplate.js
@@ -194,8 +194,9 @@ class PlotBoilerplate {
             draggable: true,
             visible: true
         };
-        if (typeof config.canvas == 'undefined')
+        if (typeof config.canvas == 'undefined') {
             throw "No canvas specified.";
+        }
         /**
          * A global config that's attached to the dat.gui control interface.
          *
@@ -775,16 +776,17 @@ class PlotBoilerplate {
         var gSize = { width: this.grid.size.x * gScale.x, height: this.grid.size.y * gScale.y };
         var cs = { width: this.canvasSize.width / 2, height: this.canvasSize.height / 2 };
         var offset = this.draw.offset.clone().inv();
+        // console.log( "drawGrid", gScale, gSize, cs, offset );
         offset.x = (Math.round(offset.x + cs.width) / Math.round(gSize.width)) * (gSize.width) / this.draw.scale.x + (((this.draw.offset.x - cs.width) / this.draw.scale.x) % gSize.width);
         offset.y = (Math.round(offset.y + cs.height) / Math.round(gSize.height)) * (gSize.height) / this.draw.scale.y + (((this.draw.offset.y - cs.height) / this.draw.scale.x) % gSize.height);
         if (this.drawConfig.drawGrid) {
             if (this.config.rasterGrid) { // TODO: move config member to drawConfig
                 draw.setCurrentId('raster');
-                draw.raster(offset, (this.canvasSize.width) / this.draw.scale.x, (this.canvasSize.height) / this.draw.scale.y, gSize.width, gSize.height, 'rgba(0,128,255,0.125)');
+                draw.raster(offset, this.canvasSize.width / this.draw.scale.x, (this.canvasSize.height) / this.draw.scale.y, gSize.width, gSize.height, 'rgba(0,128,255,0.125)');
             }
             else {
                 draw.setCurrentId('grid');
-                draw.grid(offset, (this.canvasSize.width) / this.draw.scale.x, (this.canvasSize.height) / this.draw.scale.y, gSize.width, gSize.height, 'rgba(0,128,255,0.095)');
+                draw.grid(offset, this.canvasSize.width / this.draw.scale.x, (this.canvasSize.height) / this.draw.scale.y, gSize.width, gSize.height, 'rgba(0,128,255,0.095)');
             }
         }
     }
@@ -1340,9 +1342,6 @@ class PlotBoilerplate {
      * @return {XYCoords} A simple object <pre>{ x : Number, y : Number }</pre> with the transformed coordinates.
      **/
     transformMousePosition(x, y) {
-        /* return { x : (x/this.config.cssScaleX-this.config.offsetX)/(this.config.scaleX),
-           y : (y/this.config.cssScaleY-this.config.offsetY)/(this.config.scaleY) }; */
-        // console.log('offset', this.config.offsetX, this.config.offsetY, this.draw.offset );
         return { x: (x / this.config.cssScaleX - this.config.offsetX) / (this.config.scaleX),
             y: (y / this.config.cssScaleY - this.config.offsetY) / (this.config.scaleY) };
     }