+++ changelog.md
@@ -1,6 +1,8 @@
 
 
 # Changelog
+* 2021-02-19
+  * Added the static helper function `drawutilssvg.transformPathData(...)` for svg path transformations (scale and translate).
 * 2021-02-18
   * Adding `PlotBoilerplate.adjustOffset(boolean)` function.
 * 2021-02-08